requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-core
faiss-cpu
pypdf
google-api-python-client
python-dotenv
